requirements.txt
nova-act
python-dotenv 